+++ mitra_activitypub/src/forwarder.rs
@@ -12,7 +12,7 @@ use apx_sdk::{
     deserialization::deserialize_into_id_array,
     utils::{
         get_core_type,
-        is_verification_method,
+        is_key_like,
     },
 };
 use serde::Deserialize;
@@ -71,7 +71,7 @@ pub fn verify_public_keys(
     for object in objects {
         // WARNING: this is not reliable if JSON-LD is used
         // https://codeberg.org/fediverse/fep/src/commit/8862845a2b71a32e254932757ef7696b6714739d/fep/2277/fep-2277.md#json-ld
-        if !is_verification_method(object) {
+        if !is_key_like(object) {
             continue;
         };
         let Some(object_id) = get_object_id_opt(object) else {