+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Treat public keys and verification methods as different core types.
+
 ## [4.12.0] - 2025-10-26
 
 ### Added